basic functionalities done for csv to json
