+++ src/main/java/org/wso2/carbon/module/csv/enums/EmptyCsvValueType.java
@@ -1,0 +1,6 @@
+package org.wso2.carbon.module.csv.enums;
+
+public enum EmptyCsvValueType {
+    NULL,
+    EMPTY
+}
